feat(commission): refactor commission handling by reorganizing controllers, middlewares, and models; add services/products commission processing
